# Improve performance of average snapshots