--- a/contrib/standalone/qasm_simulator.cpp
+++ b/contrib/standalone/qasm_simulator.cpp
@@ -142,7 +142,8 @@ int main(int argc, char **argv) {
 
     // Initialize simulator
     AER::Simulator::QasmController sim;
-    out << sim.execute(qobj).to_json().dump(4) << std::endl;
+    auto result = sim.execute(qobj).to_json();
+    out << result.dump(4) << std::endl;
 
     // Check if execution was successful.
     bool success = false;