--- a/releasenotes/notes/improve-ave-snapshots-768767618ff00c2d.yaml
+++ b/releasenotes/notes/improve-ave-snapshots-768767618ff00c2d.yaml
@@ -1,0 +1,4 @@
+---
+upgrade:
+  - |
+    Improve the performance of average snapshot data in simulator results.