<!--
⚠️ If you do not respect this template, your pull request will be closed.
⚠️ Your pull request title should be short detailed and understandable for all.
⚠️ Also, please add it in the CHANGELOG file under Unreleased section.
⚠️ If your pull request fixes an open issue, please link to the issue.

✅ I have added the tests to cover my changes.
✅ I have updated the documentation accordingly.
✅ I have read the CONTRIBUTING document.
-->

### Summary

Previously the mean and variance were returned as a copy of stored
accumulated data. This allows them to modify the stored data and return
a reference that can be moved into result containers which can help
avoid copies for large data objects like probability and density matrix
snapshots. To make this work it was also necessary to allow JSON conversion of
results to be non const.

### Details and comments

Memory profiler comparisons for following circuit:

```py
nq = 14
circ = QuantumCircuit(nq)
circ.snapshot_density_matrix('state')

backend = QasmSimulator()
backend_options = {
    "method": "density_matrix",
    "fusion_enabled": False,
    "truncate_enabled": False,
    "max_parallel_threads": 1
}
qobj = assemble(circ, shots=1)

start = time.time()
result = backend.run(qobj, backend_options=backend_options).result()
stop = time.time()
time_py = stop - start
time_cpp = result.metadata['time_taken']
```

For this PR: time_py = 29.2 s, time_cpp = 26.7 s
For master: time_py = 31.8 s, time_cpp = 58.3 s

And memory profiling PR:

![mprof_dm_14q_pr](https://user-images.githubusercontent.com/2235104/85475412-0efd1900-b584-11ea-96c9-b0f92b4ecf70.png)

vs Master:

![mprof_dm_14q_master](https://user-images.githubusercontent.com/2235104/85475433-158b9080-b584-11ea-933a-fd3038395fdf.png)